alteracoes e implementacoes

+++ payment-system/payment-system/src/main/java/com/esther/payment_system/service/contract/RefundService.java
@@ -2,7 +2,9 @@ package com.esther.payment_system.service.contract;
 
 import com.esther.payment_system.entity.Refund;
 
+import java.math.BigDecimal;
+
 public interface RefundService {
 
-    Refund processRefund (Long paymentId, Double amount, String reason);
+    Refund processRefund (Long paymentId, BigDecimal amount, String reason);
 }
